+++ lib/acts_as_tenant/sidekiq.rb
@@ -2,11 +2,13 @@ module ActsAsTenant::Sidekiq
   # Get the current tenant and store in the message to be sent to Sidekiq.
   class Client
     def call(worker_class, msg, queue, redis_pool)
-      msg['acts_as_tenant'] ||=
-        {
-            'class' => ActsAsTenant.current_tenant.class.name,
-            'id' => ActsAsTenant.current_tenant.id
-        } if ActsAsTenant.current_tenant.present?
+      if ActsAsTenant.current_tenant.present?
+        msg["acts_as_tenant"] ||=
+          {
+            "class" => ActsAsTenant.current_tenant.class.name,
+            "id" => ActsAsTenant.current_tenant.id
+          }
+      end
 
       yield
     end
@@ -15,8 +17,8 @@ module ActsAsTenant::Sidekiq
   # Pull the tenant out and run the current thread with it.
   class Server
     def call(worker_class, msg, queue)
-      if msg.has_key?('acts_as_tenant')
-        account = msg['acts_as_tenant']['class'].constantize.find msg['acts_as_tenant']['id']
+      if msg.has_key?("acts_as_tenant")
+        account = msg["acts_as_tenant"]["class"].constantize.find msg["acts_as_tenant"]["id"]
         ActsAsTenant.with_tenant account do
           yield
         end