+++ spec/active_record_helper.rb
@@ -1,10 +1,10 @@
-require 'rails/all'
-require 'database_cleaner'
-require 'yaml'
+require "rails/all"
+require "database_cleaner"
+require "yaml"
 
-dbconfig = YAML::load(IO.read(File.join(File.dirname(__FILE__), 'database.yml')))
+dbconfig = YAML.safe_load(IO.read(File.join(File.dirname(__FILE__), "database.yml")))
 ActiveRecord::Base.logger = Logger.new(File.join(File.dirname(__FILE__), "debug.log"))
-ActiveRecord::Base.establish_connection(dbconfig[ENV['DB'] || 'sqlite'])
+ActiveRecord::Base.establish_connection(dbconfig[ENV["DB"] || "sqlite"])
 
 RSpec.configure do |config|
   config.before(:suite) do