+++ spec/acts_as_tenant/sidekiq_spec.rb
@@ -1,28 +1,28 @@
-require 'spec_helper'
-require 'sidekiq'
-require 'acts_as_tenant/sidekiq'
+require "spec_helper"
+require "sidekiq"
+require "acts_as_tenant/sidekiq"
 
 describe ActsAsTenant::Sidekiq do
   after { ActsAsTenant.current_tenant = nil }
   let(:account) { Account.new(id: 1234) }
-  let(:message) { { 'acts_as_tenant' => { 'class' => 'Account', 'id' => 1234 } } }
+  let(:message) { {"acts_as_tenant" => {"class" => "Account", "id" => 1234}} }
 
   describe ActsAsTenant::Sidekiq::Client do
     subject { ActsAsTenant::Sidekiq::Client.new }
 
-    it 'saves tenant if present' do
+    it "saves tenant if present" do
       ActsAsTenant.current_tenant = account
 
       msg = {}
-      subject.call(nil, msg, nil, nil) { }
+      subject.call(nil, msg, nil, nil) {}
       expect(msg).to eq message
     end
 
-    it 'does not set tenant if not present' do
+    it "does not set tenant if not present" do
       expect(ActsAsTenant.current_tenant).to be_nil
 
       msg = {}
-      subject.call(nil, msg, nil, nil) { }
+      subject.call(nil, msg, nil, nil) {}
       expect(msg).not_to eq message
     end
   end
@@ -30,7 +30,7 @@ describe ActsAsTenant::Sidekiq do
   describe ActsAsTenant::Sidekiq::Server do
     subject { ActsAsTenant::Sidekiq::Server.new }
 
-    it 'restores tenant if tenant saved' do
+    it "restores tenant if tenant saved" do
       expect(Account).to receive(:find).with(1234).once { account }
 
       msg = message
@@ -40,7 +40,7 @@ describe ActsAsTenant::Sidekiq do
       expect(ActsAsTenant.current_tenant).to be_nil
     end
 
-    it 'runs without tenant if no tenant saved' do
+    it "runs without tenant if no tenant saved" do
       expect(Account).not_to receive(:find)
 
       msg = {}
@@ -51,9 +51,9 @@ describe ActsAsTenant::Sidekiq do
     end
   end
 
-  describe 'Sidekiq configuration' do
-    describe 'client configuration' do
-      it 'includes ActsAsTenant client' do
+  describe "Sidekiq configuration" do
+    describe "client configuration" do
+      it "includes ActsAsTenant client" do
         expect(Sidekiq.client_middleware.exists?(ActsAsTenant::Sidekiq::Client)).to eq(true)
       end
     end