+++ spec/acts_as_tenant/tenant_by_filter_spec.rb
@@ -1,6 +1,6 @@
 require "spec_helper"
 
-#Setup test specific ApplicationController
+# Setup test specific ApplicationController
 class Account
   attr_accessor :name
 end
@@ -12,22 +12,21 @@ class ApplicationController2 < ActionController::Base
 
   def your_method_that_finds_the_current_tenant
     current_account = Account.new
-    current_account.name = 'account1'
+    current_account.name = "account1"
     set_current_tenant(current_account)
   end
-
 end
 
 # Start testing
-describe ApplicationController2, :type => :controller do
+describe ApplicationController2, type: :controller do
   controller do
     def index
-      render :plain => "custom called"
+      render plain: "custom called"
     end
   end
 
-  it 'Finds the correct tenant using the filter command' do
+  it "Finds the correct tenant using the filter command" do
     get :index
-    expect(ActsAsTenant.current_tenant.name).to eq 'account1'
+    expect(ActsAsTenant.current_tenant.name).to eq "account1"
   end
 end