+++ spec/acts_as_tenant/test_tenant_middleware_spec.rb
@@ -1,10 +1,31 @@
-require 'spec_helper'
-require 'acts_as_tenant/test_tenant_middleware'
-require 'active_record_models'
+require "spec_helper"
+require "acts_as_tenant/test_tenant_middleware"
+require "active_record_models"
+
+class TestRackApp1
+  def call(_env)
+    ActsAsTenant.current_tenant = Account.first
+    TestReceiver.assert_current_id(ActsAsTenant.current_tenant.id)
+    ActsAsTenant.current_tenant = nil
+    [200, {}, ["OK"]]
+  end
+end
+
+class TestRackApp2
+  def call(_env)
+    TestReceiver.assert_current_id(ActsAsTenant.current_tenant.try(:id))
+    [200, {}, ["OK"]]
+  end
+end
+
+class TestReceiver
+  def self.assert_current_id(id)
+  end
+end
 
 describe ActsAsTenant::TestTenantMiddleware do
   after { ActsAsTenant.current_tenant = nil }
-  subject { request.get('/some/path') }
+  subject { request.get("/some/path") }
 
   let(:middleware) { described_class.new(app) }
   let(:request) { Rack::MockRequest.new(middleware) }
@@ -12,33 +33,13 @@ describe ActsAsTenant::TestTenantMiddleware do
   let!(:account1) { Account.create }
   let!(:account2) { Account.create }
 
-  class TestRackApp1
-    def call(_env)
-      ActsAsTenant.current_tenant = Account.first
-      TestReceiver.assert_current_id(ActsAsTenant.current_tenant.id)
-      ActsAsTenant.current_tenant = nil
-      [200, {}, ['OK']]
-    end
-  end
-
-  class TestRackApp2
-    def call(_env)
-      TestReceiver.assert_current_id(ActsAsTenant.current_tenant.try(:id))
-      [200, {}, ['OK']]
-    end
-  end
-
-  class TestReceiver
-    def self.assert_current_id(id); end
-  end
-
-  context 'when test_tenant is nil before processing' do
+  context "when test_tenant is nil before processing" do
     before { ActsAsTenant.test_tenant = nil }
 
-    context 'that switches tenancies' do
+    context "that switches tenancies" do
       let(:app) { TestRackApp1.new }
 
-      it 'should remain nil after processing' do
+      it "should remain nil after processing" do
         expect(ActsAsTenant.current_tenant).to be_nil
         expect(TestReceiver).to receive(:assert_current_id).with(account1.id)
         expect(subject.status).to eq 200
@@ -46,10 +47,10 @@ describe ActsAsTenant::TestTenantMiddleware do
       end
     end
 
-    context 'that does not switch tenancies' do
+    context "that does not switch tenancies" do
       let(:app) { TestRackApp2.new }
 
-      it 'should remain nil after processing' do
+      it "should remain nil after processing" do
         expect(ActsAsTenant.current_tenant).to be_nil
         expect(TestReceiver).to receive(:assert_current_id).with(nil)
         expect(subject.status).to eq 200
@@ -58,13 +59,13 @@ describe ActsAsTenant::TestTenantMiddleware do
     end
   end
 
-  context 'when test_tenant is assigned before processing' do
+  context "when test_tenant is assigned before processing" do
     before { ActsAsTenant.test_tenant = account2 }
 
-    context 'that switches tenancies' do
+    context "that switches tenancies" do
       let(:app) { TestRackApp1.new }
 
-      it 'should remain assigned after processing' do
+      it "should remain assigned after processing" do
         expect(ActsAsTenant.current_tenant).to eq account2
         expect(TestReceiver).to receive(:assert_current_id).with(account1.id)
         expect(subject.status).to eq 200
@@ -72,10 +73,10 @@ describe ActsAsTenant::TestTenantMiddleware do
       end
     end
 
-    context 'that does not switch tenancies' do
+    context "that does not switch tenancies" do
       let(:app) { TestRackApp2.new }
 
-      it 'should remain assigned after processing' do
+      it "should remain assigned after processing" do
         expect(ActsAsTenant.current_tenant).to eq account2
         expect(TestReceiver).to receive(:assert_current_id).with(nil)
         expect(subject.status).to eq 200