+++ spec/spec_helper.rb
@@ -1,10 +1,17 @@
-$LOAD_PATH.unshift(File.join(File.dirname(__FILE__), '..', 'lib'))
+$LOAD_PATH.unshift(File.join(File.dirname(__FILE__), "..", "lib"))
 $LOAD_PATH.unshift(File.dirname(__FILE__))
 
 require "active_record_helper"
 
-require 'rspec/rails'
-require 'acts_as_tenant'
+# Setup a test app
+module Rollcall
+  class Application < Rails::Application; end
+end
+
+Rollcall::Application.config.secret_key_base = "1234567890123456789012345678901234567890"
+
+require "rspec/rails"
+require "acts_as_tenant"
 
 RSpec.configure do |config|
   config.after(:each) do
@@ -13,11 +20,3 @@ RSpec.configure do |config|
 
   config.infer_base_class_for_anonymous_controllers = true
 end
-
-# Setup a test app
-module Rollcall
-  class Application < Rails::Application; end
-end
-
-Rollcall::Application.config.secret_token = '1234567890123456789012345678901234567890'
-Rollcall::Application.config.secret_key_base = '1234567890123456789012345678901234567890'